ProiectJava/src/main/java/com/unibuc/ro/controller/ClientController.java
package com.unibuc.ro.controller;

import com.unibuc.ro.model.Client;
import com.unibuc.ro.service.ClientService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.net.URI;

@RestController
@RequestMapping("/client")
public class ClientController {
    private final ClientService clientService;
    @Autowired
    public ClientController(ClientService clientService) {
        this.clientService = clientService;
    }
    @PostMapping
    public ResponseEntity<Client> addClient(@RequestBody Client client) {
        return ResponseEntity.created(URI.create("/"+client.getId())).body(clientService.save(client));
    }
    @GetMapping("/{id}")
    public ResponseEntity<Client> findById(@PathVariable Long id) {
        return ResponseEntity.ok().body(clientService.findById(id));
    }
}
